+++ app/Modules/Layouts/Views/layouts/_js_global.blade.php
@@ -1,10 +1,3 @@
-{{--for bootstrap 3--}}
-{{--<style>--}}
-{{--.swal2-popup {--}}
-{{--font-size: 1.6rem !important;--}}
-{{--}--}}
-{{--</style>--}}
-
 <script type="text/javascript">
 
     function notify(message, type) {