+++ resources/js/app.js
@@ -11,17 +11,35 @@ import 'jquery-ui/ui/widgets/autocomplete';
 import 'jquery-ui/ui/widgets/dialog';
 import 'jquery-ui/ui/widgets/sortable';
 import 'bootstrap-switch-button/dist/bootstrap-switch-button.min';
+//xdsoft datetimepicker
+//make sure to use from the BUILD directory...
+import 'jquery-datetimepicker/build/jquery.datetimepicker.full.min';
+//datatables.net
+import 'datatables.net/js/jquery.dataTables.min';
+import 'datatables.net-bs4/js/dataTables.bootstrap4.min'
 
-//window.Vue = require('vue');
+window.Vue = require('vue');
+
+/**
+ * The following block of code may be used to automatically register your
+ * Vue components. It will recursively scan this directory for the Vue
+ * components and automatically register them with their "basename".
+ *
+ * Eg. ./components/ExampleComponent.vue -> <example-component></example-component>
+ */
+
+// const files = require.context('./', true, /\.vue$/i)
+// files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
+
+// Vue.component('example-component', require('./components/ExampleComponent.vue').default);
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
  * the page. Then, you may begin adding components to this application
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
-
-//Vue.component('example-component', require('./components/ExampleComponent.vue'));
+// Vue.config.productionTip = false;
 
 // const app = new Vue({
-//     el: '#app'
+//     el: '#app',
 // });